+++ Tests/GRDBTests/PrimaryKeyInfoTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-import GRDB
+@testable import GRDB
 
 class PrimaryKeyInfoTests: GRDBTestCase {
     
@@ -11,6 +11,7 @@ class PrimaryKeyInfoTests: GRDBTestCase {
             XCTAssertEqual(primaryKey.columns, [Column.rowID.name])
             XCTAssertNil(primaryKey.rowIDColumn)
             XCTAssertTrue(primaryKey.isRowID)
+            XCTAssertTrue(primaryKey.tableHasRowID)
         }
     }
     
@@ -22,6 +23,7 @@ class PrimaryKeyInfoTests: GRDBTestCase {
             XCTAssertEqual(primaryKey.columns, ["id"])
             XCTAssertEqual(primaryKey.rowIDColumn, "id")
             XCTAssertTrue(primaryKey.isRowID)
+            XCTAssertTrue(primaryKey.tableHasRowID)
         }
     }
     
@@ -33,6 +35,7 @@ class PrimaryKeyInfoTests: GRDBTestCase {
             XCTAssertEqual(primaryKey.columns, ["name"])
             XCTAssertNil(primaryKey.rowIDColumn)
             XCTAssertFalse(primaryKey.isRowID)
+            XCTAssertTrue(primaryKey.tableHasRowID)
         }
     }
     
@@ -44,6 +47,31 @@ class PrimaryKeyInfoTests: GRDBTestCase {
             XCTAssertEqual(primaryKey.columns, ["a", "b"])
             XCTAssertNil(primaryKey.rowIDColumn)
             XCTAssertFalse(primaryKey.isRowID)
+            XCTAssertTrue(primaryKey.tableHasRowID)
+        }
+    }
+    
+    func testNonRowIDPrimaryKeyWithoutRowID() throws {
+        let dbQueue = try makeDatabaseQueue()
+        try dbQueue.inDatabase { db in
+            try db.execute(sql: "CREATE TABLE items (name TEXT PRIMARY KEY) WITHOUT ROWID")
+            let primaryKey = try db.primaryKey("items")
+            XCTAssertEqual(primaryKey.columns, ["name"])
+            XCTAssertNil(primaryKey.rowIDColumn)
+            XCTAssertFalse(primaryKey.isRowID)
+            XCTAssertFalse(primaryKey.tableHasRowID)
+        }
+    }
+    
+    func testCompoundPrimaryKeyWithoutRowID() throws {
+        let dbQueue = try makeDatabaseQueue()
+        try dbQueue.inDatabase { db in
+            try db.execute(sql: "CREATE TABLE items (a TEXT, b INTEGER, PRIMARY KEY (a,b)) WITHOUT ROWID")
+            let primaryKey = try db.primaryKey("items")
+            XCTAssertEqual(primaryKey.columns, ["a", "b"])
+            XCTAssertNil(primaryKey.rowIDColumn)
+            XCTAssertFalse(primaryKey.isRowID)
+            XCTAssertFalse(primaryKey.tableHasRowID)
         }
     }
 }