+++ Tests/GRDBTests/RecordPrimaryKeyHiddenRowIDTests.swift
@@ -477,17 +477,6 @@ class RecordPrimaryKeyHiddenRowIDTests : GRDBTestCase {
     }
     
     
-    // MARK: - Select Primary Key
-    
-    func testSelectPrimaryKey() throws {
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.inDatabase { db in
-            let request: QueryInterfaceRequest<Int64> = Person.selectPrimaryKey()
-            try assertEqualSQL(db, request, "SELECT \"rowid\" FROM \"persons\"")
-        }
-    }
-    
-    
     // MARK: - Fetch With Primary Key
     
     func testFetchCursorWithPrimaryKeys() throws {