+++ Tests/GRDBTests/RecordPrimaryKeyRowIDTests.swift
@@ -473,17 +473,6 @@ class RecordPrimaryKeyRowIDTests: GRDBTestCase {
     }
     
     
-    // MARK: - Select Primary Key
-    
-    func testSelectPrimaryKey() throws {
-        let dbQueue = try makeDatabaseQueue()
-        try dbQueue.inDatabase { db in
-            let request: QueryInterfaceRequest<Int64> = Person.selectPrimaryKey()
-            try assertEqualSQL(db, request, "SELECT \"id\" FROM \"persons\"")
-        }
-    }
-    
-    
     // MARK: - Fetch With Primary Key
     
     func testFetchCursorWithPrimaryKeys() throws {